Finished the website
Coded responsiveness of everything that was needed in the website. Also fixed the images that appeared cropped out. However, Images are still high resolution, it would be better if I had better pictures first so that I can use those in the website. Then I can lower the resolution so that loading times are low. Maybe adding a loading page would be nice as well.

Major things left:
-Add payment gateway
-Deploy Website to the internet